tests/requirements-dev.txt
Jinja2>=3.1,<4
PyYAML>=6,<7